+++ packages/vendure-plugin-metrics/test/metrics.spec.ts
@@ -1,3 +1,4 @@
+import { DefaultLogger, LogLevel, mergeConfig } from '@vendure/core';
 import {
   createTestEnvironment,
   registerInitializer,
@@ -6,21 +7,17 @@ import {
   testConfig,
   TestServer,
 } from '@vendure/testing';
-import { DefaultLogger, LogLevel, mergeConfig } from '@vendure/core';
-import { testPaymentMethod } from '../../test/src/test-payment-method';
+import { afterAll, beforeAll, describe, expect, it } from 'vitest';
+import getFilesInAdminUiFolder from '../../test/src/compile-admin-ui.util';
 import { initialData } from '../../test/src/initial-data';
+import { createSettledOrder } from '../../test/src/shop-utils';
+import { testPaymentMethod } from '../../test/src/test-payment-method';
 import {
+  AdvancedMetricSummariesQuery,
   AdvancedMetricSummary,
   MetricsPlugin,
-  AdvancedMetricSummariesQuery,
 } from '../src';
 import { GET_METRICS } from '../src/ui/queries.graphql';
-import { expect, describe, beforeAll, afterAll, it, vi, test } from 'vitest';
-import { createSettledOrder } from '../../test/src/shop-utils';
-import path from 'path';
-import * as fs from 'fs';
-import { compileUiExtensions } from '@vendure/ui-devkit/compiler';
-import getFilesInAdminUiFolder from '../../test/src/compile-admin-ui.util';
 
 describe('Metrics', () => {
   let shopClient: SimpleGraphQLClient;
@@ -78,15 +75,24 @@ describe('Metrics', () => {
     await adminClient.asSuperAdmin();
     const { advancedMetricSummaries } =
       await adminClient.query<AdvancedMetricSummariesQuery>(GET_METRICS);
-    const averageorderValue = advancedMetricSummaries.find(
+    const averageOrderValue = advancedMetricSummaries.find(
       (m) => m.code === 'aov'
+    )!;
+    const revenuePerProduct = advancedMetricSummaries.find(
+      (m) => m.code === 'revenue-per-product'
     )!;
     const salesPerProduct = advancedMetricSummaries.find(
       (m) => m.code === 'sales-per-product'
     )!;
-    expect(advancedMetricSummaries.length).toEqual(2);
-    expect(averageorderValue.series[0].values.length).toEqual(13);
-    expect(averageorderValue.labels.length).toEqual(13);
+    expect(advancedMetricSummaries.length).toEqual(3);
+    expect(averageOrderValue.series[0].values.length).toEqual(13);
+    expect(averageOrderValue.labels.length).toEqual(13);
+    // All orders are 4102 without tax, so that the AOV
+    expect(averageOrderValue.series[0].values[12]).toEqual(4102);
+    expect(revenuePerProduct.series[0].values.length).toEqual(13);
+    expect(revenuePerProduct.labels.length).toEqual(13);
+    // All orders are 4102 without tax, and we placed 3 orders
+    expect(revenuePerProduct.series[0].values[12]).toEqual(3 * 4102); //12306
     expect(salesPerProduct.series[0].values.length).toEqual(13);
     expect(salesPerProduct.labels.length).toEqual(13);
   });
@@ -100,20 +106,30 @@ describe('Metrics', () => {
     const averageorderValue = advancedMetricSummaries.find(
       (m) => m.code === 'aov'
     )!;
+    expect(advancedMetricSummaries.length).toEqual(3);
+    expect(averageorderValue.series[0].values.length).toEqual(13);
+    expect(averageorderValue.labels.length).toEqual(13);
+    const revenuePerProduct = advancedMetricSummaries.find(
+      (m) => m.code === 'revenue-per-product'
+    )!;
+    // For revenue per product we expect 2 series: one for each variant
+    expect(revenuePerProduct.series[0].values.length).toEqual(13);
+    expect(revenuePerProduct.series[1].values.length).toEqual(13);
+    expect(revenuePerProduct.labels.length).toEqual(13);
+    // Expect the first series (variant 1), to have 389700 revenue in last month
+    expect(revenuePerProduct.series[0].values[12]).toEqual(3897);
+    // Expect the first series (variant 2), to have 839400 revenue in last month
+    expect(revenuePerProduct.series[1].values[12]).toEqual(8394);
     const salesPerProduct = advancedMetricSummaries.find(
       (m) => m.code === 'sales-per-product'
     )!;
-    expect(advancedMetricSummaries.length).toEqual(2);
-    expect(advancedMetricSummaries.length).toEqual(2);
-    expect(averageorderValue.series[0].values.length).toEqual(13);
-    expect(averageorderValue.labels.length).toEqual(13);
     // For sales per product we expect 2 series: one for each variant
     expect(salesPerProduct.series[0].values.length).toEqual(13);
     expect(salesPerProduct.series[1].values.length).toEqual(13);
     expect(salesPerProduct.labels.length).toEqual(13);
-    // Expect the first series (variant 1), to have 3 sales in last month
+    // Expect the first series (variant 1), to have 3 revenue in last month
     expect(salesPerProduct.series[0].values[12]).toEqual(3);
-    // Expect the first series (variant 2), to have 6 sales in last month
+    // Expect the first series (variant 2), to have 6 revenue in last month
     expect(salesPerProduct.series[1].values[12]).toEqual(6);
   });